requirements.txt
face-recognition>=1.3.0
numpy>=1.20.2
glob2>=0.7
tensorflow>=2.4.0
pillow>=7.2.0
matplotlib>=3.4.1
tqdm>=4.59.0 